backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-socketio==5.11.4
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.12
aiofiles==24.1.0
